Move extension types to separate sub-dir
