+++ lib/specifications/Specification.js
@@ -192,13 +192,13 @@ class Specification {
 			return createAndInitializeSpec("Module", params);
 		}
 		case "task": {
-			return createAndInitializeSpec("Task", params);
+			return createAndInitializeSpec("extensions/Task", params);
 		}
 		case "server-middleware": {
-			return createAndInitializeSpec("ServerMiddleware", params);
+			return createAndInitializeSpec("extensions/ServerMiddleware", params);
 		}
 		case "project-shim": {
-			return createAndInitializeSpec("ProjectShim", params);
+			return createAndInitializeSpec("extensions/ProjectShim", params);
 		}
 		default:
 			throw new Error(
