+++ lib/specifications/types/extensions/ProjectShim.js
@@ -1,5 +1,4 @@
-// TODO: Move to "extensions" dir
-const Extension = require("../Extension");
+const Extension = require("../../Extension");
 
 class ProjectShim extends Extension {
 	constructor(parameters) {
