+++ lib/specifications/types/extensions/ServerMiddleware.js
@@ -1,4 +1,4 @@
-const Extension = require("../Extension");
+const Extension = require("../../Extension");
 
 class ServerMiddleware extends Extension {
 	constructor(parameters) {
