+++ lib/specifications/types/extensions/Task.js
@@ -1,4 +1,4 @@
-const Extension = require("../Extension");
+const Extension = require("../../Extension");
 
 class Task extends Extension {
 	constructor(parameters) {
